Converted login prologue to ViewPager2
