+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/LoginPrologueFragment.kt
@@ -45,7 +45,7 @@ class LoginPrologueFragment : Fragment(R.layout.fragment_login_prologue) {
             prologueFinishedListener?.onSecondaryButtonClicked()
         }
 
-        binding.viewPager.initViewPager(childFragmentManager)
+        binding.viewPager.adapter = LoginPrologueAdapter(this)
         binding.viewPagerIndicator.setupFromViewPager(binding.viewPager)
 
         if (savedInstanceState == null) {
